PointBlocks - Copy folder replaced. Quite some changes in the interaction file.
Interaction changed so that blocks cannot walk besides each other and be on each other if not on meeting points. To do this also created has_target function in object_classes character class returning True if the character passed has a target. Characters will not move if one of them already has a target.