First build of the level builder done - with plenty of bugs, but works! Game constant variables changed.
The constant variables have been changed, so that the game supports a bigger range of tile sizes. One problem is that the level files only support the tile size that they were originally made in. May be fixable, have to think about it.
Also, some ideas for new tile types stored in my random pics folder. Exciting as always.